requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.16.11
botocore==1.19.11
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-bootstrap4==2.3.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-sendgrid-v5==0.9.0
django-storages==1.10.1
djangorestframework==3.12.2
Faker==4.14.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.3
phonenumbers==8.12.18
Pillow==8.0.1
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
sendgrid==6.4.7
six==1.15.0
soupsieve==2.0.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
starkbank-ecdsa==1.1.0
text-unidecode==1.3
urllib3==1.25.11
whitenoise==5.2.0
wincertstore==0.2
